+++ flameboi/parse_event.py
@@ -1,0 +1,5 @@
+def get_payload(payload) -> dict:
+    event = payload.get("event", {})
+
+def build_dict(channel: str, text: str = None, subtype: str = None, user: str = None, ) -> dict:
+    print()
