+++ flameboi/routing.py
@@ -1,0 +1,52 @@
+def handle_team_join():
+    """
+    Returns the list of channels available to the bot.
+
+    :return: The list of channels as a dict.
+    :rtype: dict
+    """
+
+
+def handle_reaction_added():
+    """
+    Returns the list of channels available to the bot.
+
+    :return: The list of channels as a dict.
+    :rtype: dict
+    """    
+
+
+def handle_pin_added():
+    """
+    Returns the list of channels available to the bot.
+
+    :return: The list of channels as a dict.
+    :rtype: dict
+    """
+
+
+def handle_message():
+    """
+    Returns the list of channels available to the bot.
+
+    :return: The list of channels as a dict.
+    :rtype: dict
+    """
+
+
+def handle_channel_join():
+    """
+    Returns the list of channels available to the bot.
+
+    :return: The list of channels as a dict.
+    :rtype: dict
+    """
+
+
+def handle_app_mention():
+    """
+    Returns the list of channels available to the bot.
+
+    :return: The list of channels as a dict.
+    :rtype: dict
+    """
